This is software be test with brake is OK, on base this version , after modify this is version ,Data:2019-04-15

+++ nHIOS_PIC16F1828_Sample_pcv08.X/MachineLearing.h
@@ -22,10 +22,7 @@ uchar Machine_N_Learning(void);   //output signal
 
 void Average_First(uchar size,uchar address);
 
-void  convertDecimalToHexa(uint number);
-
-
-
+void  convertDecimalToHexa(ulong number);
 void Average_Second(uchar size_n,uchar address);
