+++ nHIOS_PIC16F1828_Sample_pcv08.X/Machinelearing.c
@@ -138,7 +138,7 @@ void Average_First(uchar size_m,uchar addr)
    *
  *********************************************************/
  
- void  convertDecimalToHexa(uint number)
+ void  convertDecimalToHexa(ulong number)
   {
 	 uchar i,index =0 ;
 	 ulong x = number;
@@ -165,5 +165,4 @@ void Average_First(uchar size_m,uchar addr)
 		delay_100us(5);
 	   
  }
-
